+++ lib/design_system/hdi/builders/details.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module DesignSystem
+  module Hdi
+    module Builders
+      # This class provides Hdi methods to display details in page content.
+      class Details < ::DesignSystem::Generic::Builders::Details
+      end
+    end
+  end
+end